requirements.txt
asgiref==3.7.2
Django==4.2.11
django-cors-headers==4.2.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.5.0
tzdata==2023.3
django-environ
django-filter
pillow
